Move smartcard state data into a separate module
At the moment, the smartcard data for the PCSC connection side and the
state data for the RDP side are bound together. This commit moves the
state data for the RDP side away from the PCSC data, to make it
possible to use the smartcard interface on the login screen (at some
point in the future).