Restructure server-to-client clipboard
In order to support HTML, we need the server-to-client clipboard
to advertise multiple formats at the same time, and be able to
respond to requests for different formats.

This commit also removes the canned responses from the advertisments
we send to the Windows side. We only now advertise formats we can
support requests for. The code to handle long or short format names
is also fixed.