Split X11-specific session code into a separate class
The monolithic session module is split into:
- A control modile (session.[hc])
- A session base class module (session_base.[hc])
- An X11 specific class, derived from the base class.

The session_parameters object gets its own include file, so that the
class files do not need to include session.h. This picks up layering
violations.

The hope is it will be simple to add another labwc-specific class
with minimal changes to other modules.